Add pdo namespace
