+++ lib/Minz/Pdo/PdoPgsql.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Minz\Pdo;
+
+class PdoPgsql extends AbstractPdo {
+	public function __construct($dsn, $username = null, $passwd = null, $options = null) {
+		parent::__construct($dsn, $username, $passwd, $options);
+		$this->exec("SET NAMES 'UTF8';");
+	}
+
+	public function dbType() {
+		return 'pgsql';
+	}
+
+	protected function preSql($statement) {
+		$statement = parent::preSql($statement);
+		return str_replace(array('`', ' LIKE '), array('"', ' ILIKE '), $statement);
+	}
+}
