+++ lib/Minz/Pdo/PdoSqlite.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Minz\Pdo;
+
+class PdoSqlite extends AbstractPdo {
+	public function __construct($dsn, $username = null, $passwd = null, $options = null) {
+		parent::__construct($dsn, $username, $passwd, $options);
+		$this->exec('PRAGMA foreign_keys = ON;');
+	}
+
+	public function dbType() {
+		return 'sqlite';
+	}
+
+	public function lastInsertId($name = null) {
+		return parent::lastInsertId();	//We discard the name, only used by PostgreSQL
+	}
+}
